+++ lib/types.d.ts
@@ -23,6 +23,7 @@ export type Config = {
         logoutRoute?: string;
         verifyEmail?: string;
         resendVerificationEmail?: string;
+        refreshToken?: string;
     };
     emailOptions?: {
         enableEmail: boolean;
@@ -34,6 +35,11 @@ export type Config = {
         verificationMailBody?: string;
         emailService?: EmailService;
     };
+    enableRefreshToken?: boolean;
+    refreshTokenOptions?: {
+        useCookie: boolean;
+        cookieOptions?: object;
+    };
 };
 export type ErrResponse = {
     code?: number;
@@ -41,7 +47,8 @@ export type ErrResponse = {
 };
 export type UpdateUser = {
     id: string | number;
-    password: string;
+    password?: string;
+    refreshToken?: string;
 };
 export type SaveUser = {
     username: string;
@@ -57,6 +64,7 @@ export type SuccessResponse = {
     [x: string]: any;
     message?: string | object;
     code?: number;
+    refreshToken?: string;
 };
 export type EmailOptions = {
     enableEmail?: boolean;