+++ src/constants.ts
@@ -11,6 +11,7 @@ export const REGISTER = '/register';
 export const LOGOUT = '/logout';
 export const VERIFY_EMAIL = '/verify-email';
 export const RESEND_VERIFICATION_EMAIL = '/resend-verification-email';
+export const REFRESH_TOKEN_ROUTE = '/refresh-token';
 
 export const REQUIRED_FIELDS = 'All fields are required!';
 export const USER_NOT_FOUND = 'User not found!';